backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-crontab==0.7.1
djangorestframework==3.14.0
pillow==10.2.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
